requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
future==0.18.3
idna==3.4
iso8601==2.0.0
pyserial==3.5
PyYAML==6.0
requests==2.31.0
urllib3==2.0.3
